+++ test/interactive/interactive_test1.cpp
@@ -723,6 +723,23 @@ TEST_F(InteractiveTest, bracketPaste2) {
   ASSERT_NO_FATAL_FAILURE(this->waitAndExpect(0, WaitStatus::EXITED, "\n"));
 }
 
+TEST_F(InteractiveTest, bracketPaste3) {
+  this->invoke("--quiet", "--norc");
+
+  ASSERT_NO_FATAL_FAILURE(this->expect(PROMPT));
+  ASSERT_NO_FATAL_FAILURE(this->changePrompt("> "));
+
+  // bracket paste with ctrl-v (insert-keycode)
+  this->send(CTRL_V);
+  this->paste("1234");
+  ASSERT_NO_FATAL_FAILURE(this->expect("> 1234"));
+  this->send("\r");
+  ASSERT_NO_FATAL_FAILURE(this->expect("\n: Int = 1234\n> "));
+
+  this->send(CTRL_D);
+  ASSERT_NO_FATAL_FAILURE(this->waitAndExpect(0, WaitStatus::EXITED, "\n"));
+}
+
 TEST_F(InteractiveTest, bracketPasteError1) {
   this->invoke("--quiet", "--norc");